server/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
Flask-Mail
Flask-RESTful
python-dotenv
sqlalchemy-serializer
PyJWT
